move the code to Expectation
