+++ great_expectations/validator/validator.py
@@ -2163,33 +2163,6 @@ set as active.
         return df.reset_index(drop=True, inplace=False)
 
     @staticmethod
-    def is_expectation_self_initializing(name: str) -> bool:
-        """
-        Given the name of an Expectation, returns a boolean that represents whether an Expectation can be auto-intialized.
-
-        Args:
-            name (str): name of Expectation
-
-        Returns:
-            boolean that represents whether an Expectation can be auto-initialized. Information also outputted to logger.
-        """
-
-        expectation_impl: "MetaExpectation" = get_expectation_impl(name)
-        if not expectation_impl:
-            raise ValidationError(
-                f"Expectation {name} was not found in the list of registered Expectations. "
-                f"Please check your configuration and try again"
-            )
-        if "auto" in expectation_impl.default_kwarg_values:
-            print(
-                f"The Expectation {name} is able to be self-initialized. Please run by using the auto=True parameter."
-            )
-            return True
-        else:
-            print(f"The Expectation {name} is not able to be self-initialized.")
-            return False
-
-    @staticmethod
     def _parse_validation_graph(
         validation_graph: ValidationGraph,
         metrics: Dict[Tuple[str, str, str], Any],
